consumer/requirements.txt
kafka-python==2.0.2
pymongo==4.1.1